requirements.txt
loguru==0.7.0
pydantic==1.10.7
python-dotenv==1.0.0
pyTelegramBotAPI==4.11.0
pafy
youtube_dl
